# apps/docs/openapi/openapi.yml
openapi: 3.0.0
info:
  title: Vantik
  description: AI First Project Management Tool
  version: 0.3.10
servers:
  - url: https://app.vantik.dev
paths:
...
  /api/v1/products:
    get:
      summary: List products
      description: >-
        What this workspace ships. A product holds no code and no issues: it
        groups the modules, and the issues hang off those.
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
        "401":
          description: Unauthorized
      tags:
        - Product
    post:
      summary: Create product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductDto"
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "400":
          description: Invalid request parameters
        "401":
          description: Unauthorized
      tags:
        - Product
  /api/v1/products/{productId}:
    post:
      summary: Update product
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProductDto"
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "401":
          description: Unauthorized
        "404":
          description: Not found
      tags:
        - Product
    delete:
      summary: Delete product
      description: Soft delete. The row stays and `deleted` is stamped.
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "401":
          description: Unauthorized
        "404":
          description: Not found
      tags:
        - Product
...
  /api/v1/modules/{moduleId}/repos/{moduleRepoId}:
    post:
      summary: Update a module repository
      parameters:
        - name: moduleId
          in: path
          required: true
          schema:
            type: string
        - name: moduleRepoId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateModuleRepoDto"
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModuleRepo"
        "401":
          description: Unauthorized
        "404":
          description: Not found
      tags:
        - Module
    delete:
      summary: Unpoint a module from a repository
      parameters:
        - name: moduleId
          in: path
          required: true
          schema:
            type: string
        - name: moduleRepoId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModuleRepo"
        "401":
          description: Unauthorized
        "404":
          description: Not found
      tags:
        - Module
  /api/v1/capabilities:
    get:
      summary: List capabilities
      description: >-
        What the software does for the people who use it, as opposed to what it
        is built from. A capability is defined by the modules that hold its
        code, and belongs to no product directly.
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Capability"
        "401":
          description: Unauthorized
      tags:
        - Capability
    post:
      summary: Create capability
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCapabilityDto"
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Capability"
        "400":
          description: Invalid request parameters
        "401":
          description: Unauthorized
      tags:
        - Capability
  /api/v1/capabilities/{capabilityId}:
    post:
      summary: Update capability
      parameters:
        - name: capabilityId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCapabilityDto"
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Capability"
        "401":
          description: Unauthorized
        "404":
          description: Not found
      tags:
        - Capability
    delete:
      summary: Delete capability
      description: Soft delete. Issues that named it keep the id.
      parameters:
        - name: capabilityId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Capability"
        "401":
          description: Unauthorized
        "404":
          description: Not found
      tags:
        - Capability
...
components:
  schemas:
...
    Product:
      type: object
      description: >-
        A thing the company ships. Before this model a team was the only thing
        that could stand for a product, and a company shipping two things needed
        two workspaces. A team is still the permission boundary, still holds the
        cycles, and still namespaces the issue numbers.
      properties:
        id:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        deleted:
          type: string
          format: date-time
          nullable: true
        name:
          type: string
        key:
          type: string
          description: A short name, for example "cloud". Unique in the workspace.
        description:
          type: string
          nullable: true
        status:
          type: string
          nullable: true
        icon:
          type: string
          nullable: true
        color:
          type: string
          nullable: true
        leadUserId:
          type: string
          nullable: true
        workspaceId:
          type: string
    CreateProductDto:
      type: object
      required:
        - name
      properties:
        name:
          type: string
        key:
          type: string
          description: The server makes one from the name if you send none.
        description:
          type: string
        status:
          type: string
        icon:
          type: string
        color:
          type: string
        leadUserId:
          type: string
    UpdateProductDto:
      type: object
      properties:
        name:
          type: string
        key:
          type: string
        description:
          type: string
        status:
          type: string
        icon:
          type: string
        color:
          type: string
        leadUserId:
          type: string
...
    ModuleRepo:
      type: object
      description: >-
        Where the code of a module is. An empty `pathPrefixes` means the module
        is the whole repository; a monorepo becomes several modules on one
        repository, each with its own prefixes.
      properties:
        id:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        deleted:
          type: string
          format: date-time
          nullable: true
        moduleId:
          type: string
        integrationAccountId:
          type: string
          nullable: true
        externalRepoId:
          type: string
        fullName:
          type: string
          description: '"owner/name" on the provider.'
        pathPrefixes:
          type: array
          items:
            type: string
        bidirectional:
          type: boolean
        isDefault:
          type: boolean
...
    UpdateModuleRepoDto:
      type: object
      properties:
        pathPrefixes:
          type: array
          items:
            type: string
        bidirectional:
          type: boolean
        isDefault:
          type: boolean
    Capability:
      type: object
      description: >-
        Something the software does for the people who use it. Defined by the
        modules that hold its code, and belonging to no product directly — one
        capability can span modules owned by several products.
      properties:
        id:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        deleted:
          type: string
          format: date-time
          nullable: true
        name:
          type: string
        description:
          type: string
          nullable: true
        status:
          type: string
          nullable: true
          description: planned, active, live or deprecated.
        moduleIds:
          type: array
          description: >-
            The modules that hold the code. An empty list means nobody has built
            it yet.
          items:
            type: string
        workspaceId:
          type: string
    CreateCapabilityDto:
      type: object
      required:
        - name
      properties:
        name:
          type: string
        description:
          type: string
        status:
          type: string
          description: Defaults to planned.
        moduleIds:
          type: array
          items:
            type: string
    UpdateCapabilityDto:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        status:
          type: string
        moduleIds:
          type: array
          items:
            type: string
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
security:
  - BearerAuth: []